HDDS-9826: fix exception handling if one Datanode is not available
