+++ hadoop-ozone/integration-test/src/test/java/org/apache/hadoop/ozone/client/rpc/TestWatchForCommit.java
@@ -248,6 +248,8 @@ public class TestWatchForCommit {
 
   @Test
   public void testWatchForCommitForRetryfailure() throws Exception {
+    GenericTestUtils.LogCapturer logCapturer =
+        GenericTestUtils.LogCapturer.captureLogs(XceiverClientRatis.LOG);
     try (XceiverClientManager clientManager = new XceiverClientManager(conf)) {
       ContainerWithPipeline container1 = storageContainerLocationClient
           .allocateContainer(HddsProtos.ReplicationType.RATIS,
@@ -267,6 +269,9 @@ public class TestWatchForCommit {
       long index = reply.getLogIndex();
       cluster.shutdownHddsDatanode(pipeline.getNodes().get(0));
       cluster.shutdownHddsDatanode(pipeline.getNodes().get(1));
+      // emulate closing pipeline when SCM detects DEAD datanodes
+      cluster.getStorageContainerManager()
+          .getPipelineManager().closePipeline(pipeline, false);
       // again write data with more than max buffer limit. This wi
       try {
         // just watch for a log index which in not updated in the commitInfo Map
@@ -282,6 +287,12 @@ public class TestWatchForCommit {
         // RuntimeException
         Assert.assertFalse(HddsClientUtils
             .checkForException(e) instanceof TimeoutException);
+        // client should not attempt to watch with
+        // MAJORITY_COMMITTED replication level, except the grpc IO issue
+        if (!logCapturer.getOutput().contains("Connection refused")) {
+          Assert.assertFalse(
+              e.getMessage().contains("Watch-MAJORITY_COMMITTED"));
+        }
       }
       clientManager.releaseClient(xceiverClient, false);
     }
@@ -369,9 +380,6 @@ public class TestWatchForCommit {
       } catch (Exception e) {
         Assert.assertTrue(HddsClientUtils
             .checkForException(e) instanceof GroupMismatchException);
-        // client should not attempt to watch with
-        // MAJORITY_COMMITTED replication level
-        Assert.assertFalse(e.getMessage().contains("Watch-MAJORITY_COMMITTED"));
       }
       clientManager.releaseClient(xceiverClient, false);
     }
